createHtmlContent for consent modal
